refactor(ui): extract RepoChipTelemetry from RepoSwitcher, drop grandfather (#855)
Move the per-repo telemetry snippet (drain inflight/queue badge + expandable
queue list, learnings, pause) and its queue-expansion state into a child;
{#key} remount replaces the parent's reset-on-repoFilter $effect. Only one
telemetry renders at a time, so the child safely owns the state.
RepoSwitcher <template> 32/63 -> 18/36; child 16/21. Both clear the Tier-1
40/60 bar, so the RepoSwitcher .fallowrc.jsonc grandfather is deleted.
